오류없게 실행되도록 수정 중(MenuScene, TownScene까지)

+++ Assets/Scripts/Managers/DataManager.cs
@@ -62,8 +62,8 @@ public class SlotMeta
 public class PlayerSummary
 {
     public int level = 1;
-    public int currentExp = 0;
-    public int maxExp = 20;
+    public float currentExp = 0;
+    public float maxExp = 20;
 
     public float spAddedStr = 0;
     public float spAddedAgi = 0;
